+++ src/api/middleware/middleware.ts
@@ -1,0 +1,5 @@
+import express from "express";
+
+export interface IMiddleware {
+    handle(req: express.Request, res: express.Response, next: express.NextFunction): any;
+}