+++ src/messageHandler/handler.ts
@@ -1,0 +1,4 @@
+import { IClient } from "../models/client";
+import { IMessage } from "../models/message";
+
+export type Handler = (client: IClient | undefined, message: IMessage) => boolean;