+++ src/messageHandler/handlers/heartbeat/index.js
@@ -1,4 +1,0 @@
-module.exports = (client) => {
-  const nowTime = new Date().getTime();
-  client.setLastPing(nowTime);
-};